refactor : article, user 리소스 폴더에 entities 폴더 생성으로 엔터티 구분
