+++ src/article/article.service.ts
@@ -1,12 +1,12 @@
 import { Injectable, Logger, NotFoundException, UnauthorizedException } from '@nestjs/common';
-import { Article } from './article.entity';
-import { ArticleStatus } from "./article-status.enum";
+import { Article } from './entities/article.entity';
+import { ArticleStatus } from "./entities/article-status.enum";
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { UpdateArticleRequestDto } from './dto/update-article-request.dto';
 import { CreateArticleRequestDto } from './dto/create-article-request.dto';
-import { User } from "src/user/user.entity";
-import { UserRole } from 'src/user/user-role.enum';
+import { User } from "src/user/entities/user.entity";
+import { UserRole } from 'src/user/entities/user-role.enum';
 
 @Injectable()
 export class ArticleService {
