+++ src/article/pipes/article-status-validation.pipe.ts
@@ -1,5 +1,5 @@
 import { PipeTransform, Injectable, ArgumentMetadata, BadRequestException } from '@nestjs/common';
-import { ArticleStatus } from '../article-status.enum';
+import { ArticleStatus } from '../entities/article-status.enum';
 
 @Injectable()
 export class ArticleStatusValidationPipe implements PipeTransform {
