+++ src/auth/custom-role.guard.ts
@@ -1,8 +1,8 @@
 import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { ROLES_KEY } from './roles.decorator';
-import { UserRole } from '../user/user-role.enum';
-import { User } from "src/user/user.entity";
+import { UserRole } from '../user/entities/user-role.enum';
+import { User } from "src/user/entities/user.entity";
 
 @Injectable()
 export class RolesGuard implements CanActivate {
