+++ src/auth/get-user.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from "@nestjs/common";
-import { User } from "src/user/user.entity";
+import { User } from "src/user/entities/user.entity";
 
 export const GetUser = createParamDecorator((data, ctx: ExecutionContext): User => {
     const req = ctx.switchToHttp().getRequest();
