+++ src/file/profilePicture/entities/profile-picture.entity.ts
@@ -1,5 +1,5 @@
 import { BaseEntity } from "src/common/base.entity";
-import { User } from "src/user/user.entity";
+import { User } from "src/user/entities/user.entity";
 import { Column, Entity, ManyToOne } from "typeorm";
 import { ProfilePictureType } from "./profile-picture-type.enum";
