+++ src/user/entities/user.entity.ts
@@ -1,5 +1,5 @@
 import { Column, Entity, OneToMany } from "typeorm";
-import { Article } from "src/article/article.entity";
+import { Article } from "src/article/entities/article.entity";
 import { UserRole } from "./user-role.enum";
 import { BaseEntity } from "src/common/base.entity";
 import { ProfilePicture } from "src/file/profilePicture/entities/profile-picture.entity";
